requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
python-dotenv==1.0.0

# Base de données
PyMySQL==1.1.0
cryptography==41.0.7

# Email (moderne)
Flask-Mailman==1.0.0

# Tâches asynchrones
Celery==5.3.4
redis==5.0.1

# Production
gunicorn==21.2.0